Move the .htaccess scanner logic into a separate class
